Can now see where passengers have booked
Customer has a new variable called seatAlloc, used to show their seat allocation

FXMLDocument's passenger table now contains a new column to show the passenger's seat allocation

Implemented a solution that updates the tableView whenever a seat is allocated or de-allocated. This updates the table's field values

Airplane.java now has new logic when assigning or cancelling seats. It will update the customer given to include a new seatAlloc, it will be a letter of the seatAllocation is
 based on the column number