Added the ability to remove passengers
FXMLDocumentController.java includes a removeCustomer method that finds a specific customer using Arrays.binarySearch(customers, customer); then only copying the data upto that index then from after the index to the end of the array.

FXMLDocumentController.java, Wario has joined the party as a placeholder! (not...Actually a big deal, just needed to add an extra to test that the delete passenger button works properly)

Customer.java, replaced getName with getAll. Now is used to compare every variable associated with this class.